backend/requirements.txt
astroid==2.4.0
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.9.8
pylint==2.5.0
python-jose-cryptodome==1.3.2
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
Flask-Cors==3.0.8